requirements.txt
tensorflow-gpu==1.4.0
keras==2.2.2
pydot
h5py==2.10.0
scikit-image
numpy
simpleitk
photutils
tifffile
libtiff
tqdm
matplotlib
opencv-python
pandas
Pillow
albumentations